Move fetchRandomStoryId to tests/ folder
