+++ src/hackerNews/tests/testHnStory.ts
@@ -1,7 +1,7 @@
 import { fetchHnStory } from "../fetchHnStory";
 import { fetchPageInfo } from "../fetchPageInfo/fetchPageInfo";
-import { fetchRandomStoryId } from "../fetchRandomStoryId";
 import { summarize } from "../summarize";
+import { fetchRandomStoryId } from "./fetchRandomStoryId";
 
 const storyId = await fetchRandomStoryId();
