+++ src/hackerNews/tests/testPostingHnStory.ts
@@ -1,6 +1,6 @@
 import { testAgent } from "../../shared/testAgent";
-import { fetchRandomStoryId } from "../fetchRandomStoryId";
 import { postHnStory } from "../postHnStory";
+import { fetchRandomStoryId } from "./fetchRandomStoryId";
 
 const storyId = await fetchRandomStoryId();
 console.log(`Hacker News story ID: ${storyId}.`);
